backend/requirements.txt
# Ooink RAG System Dependencies

# Core dependencies
numpy==1.24.3
sentence-transformers==2.2.2
openai==0.28.1
python-dotenv==1.0.0

# Utilities
pathlib2==2.3.7; python_version < '3.4'
